SonLVLAPI - Make S3K collision respect BlockMax setting
Adjust the Load/SaveLevelColInds functions to make use of the current game's BlockMax setting, rather than having hardcoded limitations of 0x300 tiles. Fixes an issue where, when BlockMax was set higher than 0x300, additional blocks could be added but their collision would not save or be able to set in the first place.